billing-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
firebase-admin==6.2.0
PyJWT==2.8.0
cryptography==41.0.7

# HTTP client for external APIs
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and storage
google-cloud-firestore==2.13.1

# Payment processing
razorpay==1.4.1

# Async support
asyncio-mqtt==0.16.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2  # For testing API calls

# Environment management
python-dotenv==1.0.0